Refactor random photo GET request
